requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.3
paramiko==2.10.1
pkg_resources==0.0.0
pycparser==2.21
PyNaCl==1.5.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
typing_extensions==4.1.1
urllib3==1.26.8
zipp==3.6.0
